refactor: extract EntryMode & EntryStatus from Entry
