+++ src/enums/entry.js
@@ -1,0 +1,11 @@
+let EntryMode = {
+  DISPLAY: "DISPLAY",
+  EDIT: "EDIT",
+}
+
+let EntryStatus = {
+  OPEN: "OPEN",
+  DONE: "DONE",
+}
+
+export { EntryMode, EntryStatus }
